+++ src/main/kotlin/Menu/Menu.kt
@@ -2,5 +2,4 @@ package Menu
 
 abstract class Menu {
 
-    abstract fun goOut()
 }
